Update fetchers.py
The `fetch_file()` function would download a file from the links in the info.txt file, and to load the snirfs in there needed to be an additional step using `read_raw_snirf()`. Now if `load_raw` argument is true, the function will return a dictionary containing the loaded raw file(s) so it doesn't have to be done manually.